sender: Fix senerio 2 and fix windowing issue
If the sender recived a corrupted packet on the last packet of the file
it would resend even thought the receiver though the transfer was
complete. Assume corrupted packets on the last window increment the
counter.

Windowing wasn't working properly because an if statement should have
been a while loop to ensure that the window is always filled.

Signed-off-by: Jesse Taube <Mr.Bossman075@gmail.com>